renamed files and folder in redux folder
